Fully implemented product and combo product to making UserOrder support. But lacks good email sending

+++ Fambruhrefral/src/main/java/coms/repository/Sizerepo.java
@@ -9,6 +9,6 @@ import coms.model.cartorder.Wishlist;
 import coms.model.product.ProductSize;
 
 	@Repository
-	public interface Sizerepo extends JpaRepository<ProductSize, Integer> {
+	public interface Sizerepo extends JpaRepository<ProductSize, Long> {
 	   
 	}
